put all email traits together rather than in digipack
